committed
In this project, we’ll look at how adding more complex indexes can help (or hurt!) your database. Let’s imagine we are running an online book store and need to keep track of what books we offer. We’ll be working with a database of three tables. The books table is created from the top selling books of all time. The customers and orders tables are randomly generated.